Fixed #64
Update the 'Location of License Text' value based on the generated .LICENSE location from the '--extract_license' in the ABOUT file if none is set.

Otherwise, it'll keep the original value.

In addition, if the provided csv doesn't have the key 'Location of License Text', it will now be created.