Fixed #99
The resource_name(file_path) use the posixpath.split to split the parent dir and the filename. However, it doesn't seem to work on the "windows" path.
For instance, if I have the file_path: C:\tmp\test.ABOUT, the result of the
left, right = posixpath.split(file_path)

is

['','C:\tmp\test.ABOUT']

meaning it treat the whole "windows" path as the filename and thus give the invalid filename character.

A simple fix would be use the os.path.basename instead of the posixpath.split.