#242
 * Integrate the mapping option "partially" as it's able to process the
user define keys.
 * Previous genabout behavior is not to convert input column keys/names
to lower case. However, to be as consistence as the --mapping (which
convert the mapping key to lower case), I've changed the behavior to
convert input column keys to lower case.
 * Updated some test code

Left ToDo:
 * Keys in the input but not in the MAPPING.CONFIG are also put in the
generated ABOUT files which is incorrect. We should only put keys that
are defined as the supported keys and keys that user define in the
MAPPING.CONFIG as valid. Others should not be included.
 * It currently ONLY look for MAPPING.CONFIG. In another word, user are
not able to provide path to another MAPPING file yet.
 * Need to write some test files