#242
Implemented the `--mapping` option. The fields that are included in the
MAPPING.CONFIG will be captured. Other custom fields that do not exist
in the MAPPING.CONFIG will be ignored.

Note that the feature to point to another mapping file still hasn't been
implemented yet. In addition, test code need to be updated too.