Create a 'check' command #281
The 'check' command will only print the following severities errors:
'CRITICAL', 'ERROR', 'WARNING'.
It is using the same approach as user run the 'inventroy'. Instead of
creating the generated CSV file and an error.log, this command will not
create any files but just print the error on the screen.

Signed-off-by: Chin Yeung Li <tli@nexb.com>