Add 'about_file_path' reference in the error #284
The 'about_file_path' reference is now added along with the error
message. In addition, the previous code dedup all the same error for all
ABOUT files, such as missing essential field, which is incorrect. If 3
ABOUT files are missing the essential field, such as 'name', we should
log this 3 times with reference of which ABOUT files have the issue.

Signed-off-by: Chin Yeung Li <tli@nexb.com>