Code update/enhancement
 * The `about_resource` is no longer an AboutResourceField. Instead, the
`about_resource_path` is. The `about_resource` can be think of as a
filename of the resource while the `about_resource_path` reference where
are these resource located. For instance, the `about_resource` can be
'commons-looging.jar` and the `about_resource_path` can be:
`./production/commons-logging.jar` and `./test/commons-logging.jar`

Signed-off-by: Chin Yeung Li <tli@nexb.com>