Fixed #327
Fixed the bugs that's reported in the comment on Jun 12. It will now
report the 'no' in both `gen` and `inventory`.
After some thought, I think we should just keep the `y, yes, n, no` and
ignore everything else. Some users may use 'x' to represent False and
'tick' as True, so we cannot assume 'x' to be True.
Current behavior:
`y or yes` will consider as True
`n or no` will consider as False
'blank' will gives warning saying the field is present but empty


Signed-off-by: Chin Yeung Li <tli@nexb.com>