Add new attrib option to compute unique ABOUT entries
The new --unique-fields option is a list of field names separated by
commas. When provided the ABOUT entries provided to a custom
attribution template is dedulicated using the values of these fields
considered together a unique key. When several ABOUT entries have the
same unique key, only the first one is kept.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>