#479 - Working in Progress to work on scancode scan as the input
 * Added a license score field
 * Trying to create a custom template for the scancode input

Note that it's not working as the template is becoming so complex which
is very hard to use and understand. In addition, the the list for the
license score and license name doesn't seem to sync correctly. In order
to solve it, instead of digging deep to create a more complex code and
template, we need to use a license object to "group" all the licenses.

Signed-off-by: Chin Yeung Li <tli@nexb.com>